+++ hopsworksai/resource_backup_test.go
@@ -401,3 +401,184 @@ func TestBackupDelete_APIerror(t *testing.T) {
 	}
 	r.Apply(t, context.TODO())
 }
+
+func TestBackupCreate_pipelineInProgress(t *testing.T) {
+	t.Parallel()
+	r := test.ResourceFixture{
+		HttpOps: []test.Operation{
+			{
+				Method: http.MethodPost,
+				Path:   "/api/backups",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"backupId" : "new-backup-id-1"
+					}
+				}`,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/backups/new-backup-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "error",
+					"code": 404
+				}`,
+				RunOnlyOnce: true,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/clusters/cluster-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"cluster": {
+							"id": "cluster-id-1",
+							"backupPipelineInProgress": true
+						}
+					}
+				}`,
+				RunOnlyOnce: true,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/backups/new-backup-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"backup": {
+							"backupId" : "new-backup-id-1",
+							"backupName": "backup-1",
+							"state": "succeed"
+						}
+					}
+				}`,
+				RunOnlyOnce: true,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/backups/new-backup-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"backup": {
+							"backupId" : "new-backup-id-1",
+							"backupName": "backup-1",
+							"state": "succeed"
+						}
+					}
+				}`,
+				RunOnlyOnce: true,
+			},
+		},
+		Resource:             backupResource(),
+		OperationContextFunc: backupResource().CreateContext,
+		State: map[string]interface{}{
+			"cluster_id":  "cluster-id-1",
+			"backup_name": "backup-1",
+		},
+		ExpectId: "new-backup-id-1",
+	}
+	r.Apply(t, context.TODO())
+}
+
+func TestBackupCreate_noCluster(t *testing.T) {
+	t.Parallel()
+	r := test.ResourceFixture{
+		HttpOps: []test.Operation{
+			{
+				Method: http.MethodPost,
+				Path:   "/api/backups",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"backupId" : "new-backup-id-1"
+					}
+				}`,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/backups/new-backup-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "error",
+					"code": 404
+				}`,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/clusters/cluster-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 404
+				}`,
+			},
+		},
+		Resource:             backupResource(),
+		OperationContextFunc: backupResource().CreateContext,
+		State: map[string]interface{}{
+			"cluster_id":  "cluster-id-1",
+			"backup_name": "backup-1",
+		},
+		ExpectError: "cluster not found for cluster id cluster-id-1",
+	}
+	r.Apply(t, context.TODO())
+}
+
+func TestBackupCreate_getCluster_error(t *testing.T) {
+	t.Parallel()
+	r := test.ResourceFixture{
+		HttpOps: []test.Operation{
+			{
+				Method: http.MethodPost,
+				Path:   "/api/backups",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"backupId" : "new-backup-id-1"
+					}
+				}`,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/backups/new-backup-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "error",
+					"code": 404
+				}`,
+			},
+			{
+				Method: http.MethodGet,
+				Path:   "/api/clusters/cluster-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 400,
+					"message": "failed to get cluster info"
+				}`,
+			},
+		},
+		Resource:             backupResource(),
+		OperationContextFunc: backupResource().CreateContext,
+		State: map[string]interface{}{
+			"cluster_id":  "cluster-id-1",
+			"backup_name": "backup-1",
+		},
+		ExpectError: "failed to get cluster info",
+	}
+	r.Apply(t, context.TODO())
+}