+++ CHANGELOG.md
@@ -5,6 +5,7 @@ BREAKING CHANGES:
 NOTES:
 
 BUG FIXES:
+* resource/hopsworksai_backup: Handle get backup if backup not found and backupPipeline is InProgress
 
 ENHANCEMENTS: